убрал ошибки proptypes
